requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
elasticsearch==8.15.1
google-api-python-client==2.143.0
google-auth==2.34.0
python-dotenv==1.0.1
requests==2.32.3
pdfminer.six==20240706
pytesseract==0.3.13
Pillow==10.4.0
